+++ DifferentialSteering.h
@@ -23,7 +23,7 @@ class DifferentialSteering
         void computeMotors(int XValue, int YValue);
         int computedLeftMotor();
         int computedRightMotor();
-        byte getComputeRange();
+        int getComputeRange();
         String toString();
 };
 #endif
